+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - [#20][GL-20] - Support loading/saving banner patterns on items
 - [#30][GL-30] - Option to override max-players-per-team in the Love at First Sight scenario
     - Added the `max-players-per-team` setting to `love_at_first_sight` in `scenarios.yml`
+- [#36][GL-36] - Option to disable sneaking requirement in the Vein Miner scenario
+    - Added the `require-sneaking` setting to `vein_miner` in `scenarios.yml`
 - Added `team.message.invite-success` to `lang.yml`
 
 ### Bug fixes
@@ -23,6 +25,7 @@
 [GL-23]: https://gitlab.com/uhccore/uhccore/-/issues/23
 [GL-29]: https://gitlab.com/uhccore/uhccore/-/issues/29
 [GL-30]: https://gitlab.com/uhccore/uhccore/-/issues/30
+[GL-36]: https://gitlab.com/uhccore/uhccore/-/issues/36
 [GL-45]: https://gitlab.com/uhccore/uhccore/-/issues/45
 
 ## [1.20.3] - 2022-12-17